+++ acefile.py
@@ -55,7 +55,7 @@ notifications, refer to:
 - https://twitter.com/droethlisberger
 """
 
-__version__     = '0.6.7-dev'
+__version__     = '0.6.7'
 __author__      = 'Daniel Roethlisberger'
 __email__       = 'daniel@roe.ch'
 __copyright__   = 'Copyright 2017, Daniel Roethlisberger'