+++ NEWS.md
@@ -1,3 +1,7 @@
+### acefile 0.6.8-dev
+
+
+
 ### acefile 0.6.7 2017-08-21
 
 -   Make restoration of mtime/atime on extraction optional, default off, and